Enhance media processing by adding segment-specific embeddings, improving unsupported format handling, and enabling advanced audio pipelines with audio diarization and voice embeddings. Adjust source separation and transcription priorities, and update conditional build references for comprehensive media analysis.

+++ src/Mostlylucid.DocSummarizer.LLamaSharp/Services/LLamaSharpLlmService.cs
@@ -267,7 +267,7 @@ public sealed class LLamaSharpLlmService : ILlmService, IDisposable
             {
                 ContextSize = sentinel ? 4096u : _config.ContextSize,
                 GpuLayerCount = _config.GpuLayerCount,
-                MainGpu = (uint)_config.GpuDeviceId,
+                MainGpu = _config.GpuDeviceId,
                 BatchSize = (uint)_config.BatchSize
             };
